Major rebranding of minecraft modpacks
So... We didn't planned to make this but to make this modpack popular, we are now going with Modrinth to upload the modpack which is fully open-source.

As a result non open-source mods needs to be removed and rebranding needs to exist.

This is what the answers are. Introducing... Kosmos / Steria.